requirements.txt
# ============================================================================
# ЗАВИСИМОСТИ ПРОЕКТА IMAGE HOSTING
# ============================================================================
# 
# Этот файл содержит все Python пакеты, необходимые для работы приложения.
# Версии зафиксированы для обеспечения стабильности и воспроизводимости.
# 
# Установка: pip install -r requirements.txt
# Обновление: pip freeze > requirements.txt
# 
# ============================================================================

# ============================================================================
# ОСНОВНОЙ ВЕБ-ФРЕЙМВОРК
# ============================================================================

# Flask - микро веб-фреймворк для Python
# Версия 3.1.1 - стабильная версия с поддержкой современных возможностей
# Используется для создания веб-приложения, маршрутизации, шаблонов
Flask==3.1.1

# Werkzeug - WSGI утилиты для Flask
# Обеспечивает HTTP сервер для разработки, отладочные инструменты
# Версия совместима с Flask 3.1.1
Werkzeug==3.1.2

# ============================================================================
# ОБРАБОТКА ИЗОБРАЖЕНИЙ
# ============================================================================

# Pillow - библиотека для работы с изображениями
# Используется для:
# - Изменения размера изображений
# - Конвертации форматов
# - Создания миниатюр
# - Валидации изображений
Pillow==10.2.0

# ============================================================================
# КОНФИГУРАЦИЯ И ПЕРЕМЕННЫЕ ОКРУЖЕНИЯ
# ============================================================================

# python-dotenv - загрузка переменных окружения из .env файла
# Позволяет хранить конфигурацию отдельно от кода
# Используется для загрузки паролей, ключей, настроек БД
python-dotenv==1.1.0

# ============================================================================
# ПРОДАКШН WSGI СЕРВЕР
# ============================================================================

# Gunicorn - Python WSGI HTTP сервер для UNIX
# Используется в продакшене вместо встроенного сервера Flask
# Обеспечивает:
# - Высокую производительность
# - Поддержку множественных worker процессов
# - Стабильность под нагрузкой
gunicorn==23.0.0

# ============================================================================
# БАЗА ДАННЫХ POSTGRESQL
# ============================================================================

# psycopg2-binary - PostgreSQL адаптер для Python
# binary версия включает скомпилированные библиотеки
# Используется для:
# - Подключения к PostgreSQL
# - Выполнения SQL запросов
# - Управления транзакциями
psycopg2-binary==2.9.9

# ============================================================================
# УПРАВЛЕНИЕ СЕССИЯМИ
# ============================================================================

# Flask-Session - расширение для управления сессиями пользователей
# Обеспечивает:
# - Безопасное хранение данных сессии
# - Поддержку различных backend'ов для сессий
# - Автоматическое управление cookies
Flask-Session==0.5.0

# ============================================================================
# ВНУТРЕННИЕ ЗАВИСИМОСТИ FLASK
# ============================================================================
# 
# Эти пакеты являются зависимостями Flask и устанавливаются автоматически.
# Указаны явно для фиксации версий и обеспечения стабильности.
# 

# MarkupSafe - безопасная обработка HTML/XML разметки
# Предотвращает XSS атаки через экранирование опасных символов
markupsafe==2.1.5

# itsdangerous - криптографические подписи для данных
# Используется Flask для подписи cookies и сессий
# Обеспечивает защиту от подделки данных
itsdangerous==2.2.0

# Click - библиотека для создания CLI интерфейсов
# Используется Flask для команд управления приложением
# Позволяет создавать кастомные команды
click==8.1.7

# Jinja2 - современный шаблонизатор для Python
# Используется Flask для рендеринга HTML шаблонов
# Обеспечивает безопасность и гибкость шаблонов
jinja2==3.1.3

# Blinker - библиотека сигналов для Python
# Используется Flask для системы событий и хуков
# Позволяет подписываться на события приложения
blinker>=1.9.0

# ============================================================================
# НОВЫЕ ЗАВИСИМОСТИ ДЛЯ ОПТИМИЗАЦИИ ПРОИЗВОДИТЕЛЬНОСТИ
# ============================================================================

# Redis для кэширования и сессий
redis==4.6.0
Flask-Session==0.5.0
Flask-Caching==2.1.0

# Асинхронная обработка
celery==5.3.4
celery[redis]==5.3.4

# Мониторинг и метрики
prometheus-client==0.19.0
psutil==5.9.6

# Улучшенная обработка изображений
# Pillow-SIMD==10.0.1.post1  # Оптимизированная версия Pillow (опционально)

# Пул соединений (уже есть psycopg2, но добавим расширения)
psycopg2-pool==1.1

# ============================================================================
# СТАНДАРТНАЯ БИБЛИОТЕКА PYTHON
# ============================================================================
# 
# Следующие модули являются частью стандартной библиотеки Python
# и не требуют отдельной установки:
# 
# - hashlib: криптографические хеш-функции (для паролей)
# - logging: система логирования приложения
# - os: взаимодействие с операционной системой
# - subprocess: запуск внешних процессов
# - datetime: работа с датой и временем
# - json: обработка JSON данных
# - uuid: генерация уникальных идентификаторов
# - pathlib: современная работа с путями файлов
# ============================================================================
